requirements.txt
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
idna==2.10
loguru==0.5.3
lxml==4.6.2
MouseInfo==0.1.3
Pillow==8.1.2
PyAutoGUI==0.9.52
PyExecJS==1.5.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.3.7
PyTweening==1.0.3
requests==2.25.1
six==1.15.0
translators==4.7.16
urllib3==1.26.3
win32-setctime==1.0.3
